requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
colorama>=0.4.6